Renamed views to components and created utils file
